tools/requirements.txt
beautifulsoup4
htmlmin
rcssmin
rjsmin
Pillow
tqdm
